dashboardが検索できない、テーブルの問題を修正 - dashboardに初回アクセス時のfetchと手動fetchのクエリが混ざり検索できなくなるので、初回fetchのロジックを削除 - 代替案としてdashboardの検索フィールド開始終了に初期値を設定 - データが存在しない場合、テーブルのエラーがあったため、0番目がnullの場合の条件を追加 - 伴うテスト修正
