+++ app/hooks/useFetchRawsData.ts
@@ -1,25 +1,15 @@
 'use client';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { supabase } from '../libs/supabase';
-import useSWR, { mutate } from 'swr';
+import { mutate } from 'swr';
 import type { Database } from '../../supabase/schema';
 
 type Props = Database['public']['Tables']['raws_data']['Row'][] | null;
 
-type ConditionsArgsType = {
+export type ConditionsArgsType = {
   startDate: string;
   endDate: string;
   person?: string;
-};
-
-const getNowMonthFirstLast = () => {
-  const nowDate = new Date();
-  const nowMonthFirst = new Date(nowDate.getFullYear(), nowDate.getMonth(), 1);
-  const nowMonthLast = new Date(nowDate.getFullYear(), nowDate.getMonth() + 1, 0);
-  return {
-    startDate: nowMonthFirst.toISOString().split('T')[0],
-    endDate: nowMonthLast.toISOString().split('T')[0],
-  };
 };
 
 const commonSupabaseFetcher = supabase
@@ -46,21 +36,10 @@ const conditionsFetcher = async (args: ConditionsArgsType) => {
 
 export const useFetchRawsData = () => {
   const [rawsData, setRawsData] = useState<Props | null>(null);
-  const { startDate, endDate } = getNowMonthFirstLast();
-  const sendingData: ConditionsArgsType = {
-    person: '',
-    startDate,
-    endDate,
-  };
-
-  const { data } = useSWR('raws_data', () => conditionsFetcher({ ...sendingData }));
-  useEffect(() => {
-    setRawsData(data?.data || null);
-  }, [data]);
 
   const mutateFetch = async (args: ConditionsArgsType) => {
     const result = await mutate('raws_data_conditions', conditionsFetcher({ ...args }));
-    setRawsData(result?.data || null);
+    setRawsData(() => result?.data || null);
   };
 
   return {
